dlp/serializers.py
from rest_framework import serializers

from .models import Pattern


class PatternSerializer(serializers.ModelSerializer):
    class Meta:
        model = Pattern
        fields = ['id', 'name', 'regex_pattern']