requirements.txt
aio-pika==9.4.3
aiormq==6.8.1
asgiref==3.8.1
black==24.10.0
click==8.1.7
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
flake8==7.1.1
idna==3.10
isort==5.13.2
mccabe==0.7.0
multidict==6.1.0
mypy-extensions==1.0.0
mysqlclient==2.2.5
packaging==24.1
pamqp==3.3.0
pathspec==0.12.1
pika==1.3.2
platformdirs==4.3.6
propcache==0.2.0
pycodestyle==2.12.1
pyflakes==3.2.0
python-dotenv==1.0.1
slack_sdk==3.33.2
sqlparse==0.5.1
tomli==2.0.2
typing_extensions==4.12.2
yarl==1.17.0
